// sdk/typescript/src/api/artifact/artifact.ts
export enum ArtifactSystem {
  UNKNOWN_SYSTEM = 0,
  AARCH64_DARWIN = 1,
  AARCH64_LINUX = 2,
  X8664_DARWIN = 3,
  X8664_LINUX = 4,
  UNRECOGNIZED = -1,
}
// ...
export interface ArtifactSource {
  digest?: string | undefined;
  excludes: string[];
  includes: string[];
  name: string;
  path: string;
}

export interface ArtifactStepSecret {
  name: string;
  value: string;
}

export interface ArtifactStep {
  entrypoint?: string | undefined;
  script?: string | undefined;
  secrets: ArtifactStepSecret[];
  arguments: string[];
  artifacts: string[];
  environments: string[];
}

export interface Artifact {
  target: ArtifactSystem;
  sources: ArtifactSource[];
  steps: ArtifactStep[];
  systems: ArtifactSystem[];
  aliases: string[];
  name: string;
}
// ...
export const ArtifactSource: MessageFns<ArtifactSource> = {
  encode(message: ArtifactSource, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.digest !== undefined) {
      writer.uint32(10).string(message.digest);
    }
    for (const v of message.excludes) {
      writer.uint32(18).string(v!);
    }
    for (const v of message.includes) {
      writer.uint32(26).string(v!);
    }
    if (message.name !== "") {
      writer.uint32(34).string(message.name);
    }
    if (message.path !== "") {
      writer.uint32(42).string(message.path);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): ArtifactSource {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseArtifactSource();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.digest = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.excludes.push(reader.string());
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.includes.push(reader.string());
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.name = reader.string();
          continue;
        }
        case 5: {
          if (tag !== 42) {
            break;
          }

          message.path = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): ArtifactSource {
    return {
      digest: isSet(object.digest) ? globalThis.String(object.digest) : undefined,
      excludes: globalThis.Array.isArray(object?.excludes) ? object.excludes.map((e: any) => globalThis.String(e)) : [],
      includes: globalThis.Array.isArray(object?.includes) ? object.includes.map((e: any) => globalThis.String(e)) : [],
      name: isSet(object.name) ? globalThis.String(object.name) : "",
      path: isSet(object.path) ? globalThis.String(object.path) : "",
    };
  },

  toJSON(message: ArtifactSource): unknown {
    const obj: any = {};
    if (message.digest !== undefined) {
      obj.digest = message.digest;
    }
    if (message.excludes?.length) {
      obj.excludes = message.excludes;
    }
    if (message.includes?.length) {
      obj.includes = message.includes;
    }
    if (message.name !== "") {
      obj.name = message.name;
    }
    if (message.path !== "") {
      obj.path = message.path;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<ArtifactSource>, I>>(base?: I): ArtifactSource {
    return ArtifactSource.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<ArtifactSource>, I>>(object: I): ArtifactSource {
    const message = createBaseArtifactSource();
    message.digest = object.digest ?? undefined;
    message.excludes = object.excludes?.map((e) => e) || [];
    message.includes = object.includes?.map((e) => e) || [];
    message.name = object.name ?? "";
    message.path = object.path ?? "";
    return message;
  },
};

function createBaseArtifactStepSecret(): ArtifactStepSecret {
  return { name: "", value: "" };
}

export const ArtifactStepSecret: MessageFns<ArtifactStepSecret> = {
  encode(message: ArtifactStepSecret, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.name !== "") {
      writer.uint32(10).string(message.name);
    }
    if (message.value !== "") {
      writer.uint32(18).string(message.value);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): ArtifactStepSecret {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseArtifactStepSecret();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.name = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.value = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): ArtifactStepSecret {
    return {
      name: isSet(object.name) ? globalThis.String(object.name) : "",
      value: isSet(object.value) ? globalThis.String(object.value) : "",
    };
  },

  toJSON(message: ArtifactStepSecret): unknown {
    const obj: any = {};
    if (message.name !== "") {
      obj.name = message.name;
    }
    if (message.value !== "") {
      obj.value = message.value;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<ArtifactStepSecret>, I>>(base?: I): ArtifactStepSecret {
    return ArtifactStepSecret.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<ArtifactStepSecret>, I>>(object: I): ArtifactStepSecret {
    const message = createBaseArtifactStepSecret();
    message.name = object.name ?? "";
    message.value = object.value ?? "";
    return message;
  },
};
// ...
export const Artifact: MessageFns<Artifact> = {
  encode(message: Artifact, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.target !== 0) {
      writer.uint32(8).int32(message.target);
    }
    for (const v of message.sources) {
      ArtifactSource.encode(v!, writer.uint32(18).fork()).join();
    }
    for (const v of message.steps) {
      ArtifactStep.encode(v!, writer.uint32(26).fork()).join();
    }
    writer.uint32(34).fork();
    for (const v of message.systems) {
      writer.int32(v);
    }
    writer.join();
    for (const v of message.aliases) {
      writer.uint32(42).string(v!);
    }
    if (message.name !== "") {
      writer.uint32(50).string(message.name);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): Artifact {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseArtifact();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.target = reader.int32() as any;
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.sources.push(ArtifactSource.decode(reader, reader.uint32()));
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.steps.push(ArtifactStep.decode(reader, reader.uint32()));
          continue;
        }
        case 4: {
          if (tag === 32) {
            message.systems.push(reader.int32() as any);

            continue;
          }

          if (tag === 34) {
            const end2 = reader.uint32() + reader.pos;
            while (reader.pos < end2) {
              message.systems.push(reader.int32() as any);
            }

            continue;
          }

          break;
        }
        case 5: {
          if (tag !== 42) {
            break;
          }

          message.aliases.push(reader.string());
          continue;
        }
        case 6: {
          if (tag !== 50) {
            break;
          }

          message.name = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): Artifact {
    return {
      target: isSet(object.target) ? artifactSystemFromJSON(object.target) : 0,
      sources: globalThis.Array.isArray(object?.sources)
        ? object.sources.map((e: any) => ArtifactSource.fromJSON(e))
        : [],
      steps: globalThis.Array.isArray(object?.steps) ? object.steps.map((e: any) => ArtifactStep.fromJSON(e)) : [],
      systems: globalThis.Array.isArray(object?.systems)
        ? object.systems.map((e: any) => artifactSystemFromJSON(e))
        : [],
      aliases: globalThis.Array.isArray(object?.aliases) ? object.aliases.map((e: any) => globalThis.String(e)) : [],
      name: isSet(object.name) ? globalThis.String(object.name) : "",
    };
  },

  toJSON(message: Artifact): unknown {
    const obj: any = {};
    if (message.target !== 0) {
      obj.target = artifactSystemToJSON(message.target);
    }
    if (message.sources?.length) {
      obj.sources = message.sources.map((e) => ArtifactSource.toJSON(e));
    }
    if (message.steps?.length) {
      obj.steps = message.steps.map((e) => ArtifactStep.toJSON(e));
    }
    if (message.systems?.length) {
      obj.systems = message.systems.map((e) => artifactSystemToJSON(e));
    }
    if (message.aliases?.length) {
      obj.aliases = message.aliases;
    }
    if (message.name !== "") {
      obj.name = message.name;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<Artifact>, I>>(base?: I): Artifact {
    return Artifact.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<Artifact>, I>>(object: I): Artifact {
    const message = createBaseArtifact();
    message.target = object.target ?? 0;
    message.sources = object.sources?.map((e) => ArtifactSource.fromPartial(e)) || [];
    message.steps = object.steps?.map((e) => ArtifactStep.fromPartial(e)) || [];
    message.systems = object.systems?.map((e) => e) || [];
    message.aliases = object.aliases?.map((e) => e) || [];
    message.name = object.name ?? "";
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends { $case: string } ? { [K in keyof Omit<T, "$case">]?: DeepPartial<T[K]> } & { $case: T["$case"] }
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}